PID: Remove QA option from pidTOF and pidTOFFull
