+++ Common/TableProducer/PID/pidTOFBase.cxx
@@ -29,7 +29,6 @@
 #include "Common/DataModel/TrackSelectionTables.h"
 #include "Common/DataModel/EventSelection.h"
 #include "Common/DataModel/FT0Corrected.h"
-#include "DPG/Tasks/AOTTrack/PID/qaPIDTOF.h"
 #include "TableHelper.h"
 #include "pidTOFBase.h"
