+++ DPG/Tasks/AOTTrack/PID/qaPIDTOFBeta.cxx
@@ -17,8 +17,13 @@
 
 #include "Framework/AnalysisTask.h"
 #include "Common/TableProducer/PID/pidTOFBase.h"
-#include "qaPIDTOF.h"
 #include "Framework/runDataProcessing.h"
+#include "Framework/HistogramRegistry.h"
+#include "Framework/StaticFor.h"
+#include "Common/DataModel/TrackSelectionTables.h"
+#include "Common/DataModel/EventSelection.h"
+#include "Common/DataModel/PIDResponse.h"
+#include "Common/DataModel/FT0Corrected.h"
 
 struct tofPidBetaQa {
   static constexpr int Np = 9;
